Separate base from subclasses
